bring become-seller in private route
